refactor: separate init and cleanup logic
The shell's initialization and cleanup logic have been refactored into
two separate functions: `init_rickshell` and `cleanup_rickshell`.

This improves code organization by encapsulating the setup and teardown
processes, making future modifications and maintenance easier. The
`init_rickshell` function handles all shell startup routines, including
initialization of necessary resources, while `cleanup_rickshell` ensures
proper resource release and cleanup when the shell exits.

Refs: #20
